feat(client): cube picker root + cube page + catalog under /[cubeId]
