+++ apps/client/src/routes/[cubeId]/+page.server.ts
@@ -1,0 +1,5 @@
+import * as repo from '$lib/server/repository';
+
+export const prerender = true;
+
+export const entries = () => repo.getCubes().map((cube) => ({ cubeId: cube.id }));
