+++ apps/client/src/routes/[cubeId]/[setId]/+page.server.ts
@@ -1,0 +1,5 @@
+import * as repo from '$lib/server/repository';
+
+export const prerender = true;
+
+export const entries = () => repo.getSets().map((set) => ({ cubeId: set.cubeId, setId: set.id }));
