+++ apps/client/src/routes/[cubeId]/[setId]/+page.svelte
