+++ apps/client/src/routes/[cubeId]/[setId]/[subsetId]/+page.server.ts
@@ -1,0 +1,12 @@
+import * as repo from '$lib/server/repository';
+
+export const prerender = true;
+
+export const entries = () =>
+	repo.getSets().flatMap((set) =>
+		repo.getSetSubsets(set.id).map((subset) => ({
+			cubeId: set.cubeId,
+			setId: set.id,
+			subsetId: subset.id
+		}))
+	);
