+++ apps/client/src/routes/[cubeId]/[setId]/[subsetId]/[caseId]/+page.svelte
