--- apps/client/src/routes/[setId]/+page.server.ts
@@ -1,5 +1,0 @@
-import * as repo from '$lib/server/repository';
-
-export const prerender = true;
-
-export const entries = () => repo.getSets().map((set) => ({ setId: set.id }));
